requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
pillow==10.4.0
PyJWT==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
razorpay==1.4.2
requests==2.32.3
setuptools==75.0.0
sqlparse==0.5.1
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.11.1
